Add tests for a bug which occurs when using small dictionaries
When using dictionaries of size up to 2**16, encoding fails
and produces an invalid file which cannot be successfully
decoded.

This adds testing of the compression -> decompression round
trip with the compression levels 1, 6 and 9.
Using compression level 1, the bug is triggered; Levels 2 to 9
seem to work fine, at least for the sample files in test/.